refactor: usecase name fix
